+++ R/aaa.R
@@ -1,0 +1,3 @@
+#' @useDynLib ggsvg, .registration=TRUE
+NULL
+